requirements.txt
folium==0.19.5
geopandas==1.0.1
matplotlib==3.10.1
numpy==2.2.5
pandas==2.2.3
plotly==6.0.1
python_dateutil==2.9.0.post0
seaborn==0.13.2
statsmodels==0.14.4
streamlit==1.44.1
streamlit_folium==0.25.0
